When cart decrement reaches 0 cart item is deleted and amount also updates according to increment and decrement clicks. I need to add this functionality to the delete button too and fix the problem leading to when cart items are 0 the amount says RNAN instead of 0

+++ Notes-App-V2/javascript/script.js
@@ -103,10 +103,16 @@ notesList.addEventListener('click', function(e){
 
         const targetNoteId = noteObjectElement.dataset.id
 
-        deleteNote(targetNoteId)
+        let deleteCheck = prompt("Are you sure you want to delete your note? enter Y/N")
 
-        // Remove from UI
-        noteObjectElement.remove()
+        if(deleteCheck == "Y" || deleteCheck == "y"){
+            deleteNote(targetNoteId)
+
+            // Remove from UI
+            noteObjectElement.remove()
+        }else{
+            return
+        }
     }
 })
